--- a/test/models/test_real_world.py
+++ b/test/models/test_real_world.py
@@ -7,14 +7,16 @@ from tinygrad.ops import GlobalCounters, LazyOp, LoadOps
 from tinygrad.ops import Device
 from tinygrad.helpers import CI, dtypes
 
+from examples.gpt2 import Transformer as GPT2Transformer, MODEL_PARAMS as GPT2_MODEL_PARAMS
 from examples.hlb_cifar10 import SpeedyResNet
-from examples.llama import Transformer, MODEL_PARAMS
+from examples.llama import Transformer as LLaMaTransformer, MODEL_PARAMS as LLAMA_MODEL_PARAMS
 from examples.stable_diffusion import UNetModel
 
 def helper_test(nm, gen, train, max_memory_allowed, max_kernels_allowed):
   tms = []
   for _ in range(4):
     GlobalCounters.reset()
+    GlobalCounters.mem_used = 0
     Device[Device.DEFAULT].synchronize()
     st = time.perf_counter_ns()
     train(*gen())
@@ -55,11 +57,25 @@ class TestRealWorld(unittest.TestCase):
     Tensor.default_type = dtypes.float16
 
     args_tiny = {"dim": 1024, "multiple_of": 256, "n_heads": 8, "n_layers": 8, "norm_eps": 1e-05, "vocab_size": 1000}
-    model = Transformer(**(args_tiny if CI else MODEL_PARAMS[1]["7B"]["args"]))
+    model = LLaMaTransformer(**(args_tiny if CI else LLAMA_MODEL_PARAMS[1]["7B"]["args"]))
     derandomize_model(model)
     @TinyJit
     def test(t): return model(t, 0).realize()
     helper_test("test_llama", lambda: (Tensor([[1,]]),), test, 0.22 if CI else 13.5, 126 if CI else 486)
+
+    Tensor.default_type = old_type
+
+  @unittest.skipUnless(Device.DEFAULT in JIT_SUPPORTED_DEVICE and Device.DEFAULT not in ["LLVM"], "needs JIT, too long on CI LLVM")
+  def test_gpt2(self):
+    old_type = Tensor.default_type
+    Tensor.default_type = dtypes.float16
+
+    args_tiny = {"dim": 1024, "n_heads": 8, "n_layers": 8, "norm_eps": 1e-5, "vocab_size": 1000}
+    model = GPT2Transformer(**(args_tiny if CI else GPT2_MODEL_PARAMS["gpt2-medium"]))
+    derandomize_model(model)
+    @TinyJit
+    def test(t): return model(t, 0).realize()
+    helper_test("test_gpt2", lambda: (Tensor([[1,]]),), test, 0.21 if CI else 0.9, 129 if CI else 369)
 
     Tensor.default_type = old_type