Move symbolic reshape and shrink inside the jitted function and jit the whole `TransformerBlock.__call__` similar to #1633. It saved about 1.5 ms python time on M1 Max with METAL. `GPU=1` is currently broken with `ValueError: probabilities contain NaN` on mac on master.

Also added a gpt2 test in `test_real_world`.

with `METAL=1 JIT=1 python examples/gpt2.py --prompt="Hello." --count=10 --temperature=0 --timing`.

this PR
```
ran model in 11.31 ms
total 43.88 ms

ran model in 11.41 ms
total 43.93 ms

ran model in 11.34 ms
total 44.10 ms
Hello. I'm a little late to the party, but
```

master
```
ran model in 12.95 ms
total 44.02 ms

ran model in 13.20 ms
total 43.97 ms

ran model in 13.11 ms
total 44.24 ms
Hello. I'm a little late to the party, but
```