@geohot 2023-09-02
Nice! It's like 3ms on my 13900K now